Changing API call logics and fixture

+++ page-fixtures/experimental-options.ts
@@ -1,10 +1,8 @@
-import {APIRequestContext, Page, test as base } from "@playwright/test";
+import {APIRequestContext, Page, TestInfo, test as base } from "@playwright/test";
 import { Utils } from "../utils/utils";
 
 export type TestOptions = {
-  UIFixture: Page;
-  APIFixture: APIRequestContext;
-  MixedFixture: string;
+  customFixture: string;
   customPage: Page;
   customRequest: APIRequestContext;
 };
@@ -13,28 +11,20 @@ let customPage: Page;
 let customRequest: APIRequestContext;
 
 export const test = base.extend<TestOptions>({
-  MixedFixture: async ({ page, request }, use) => {
-    await Utils.deleteFolder('allure-results');
-    customPage = page;
-    customRequest = request;
-    await page.goto("/");
-    await use("");
-    await page.close();
-    await Utils.deleteFolder('test-results');
-  },
-  UIFixture: async ({ page, request }, use) => {
-    await Utils.deleteFolder('allure-results');
-    customPage = page;
-    customRequest = request;
-    await page.goto("/");
-    await use(customPage);
-    await page.close();
-    await Utils.deleteFolder('test-results');
-  },
-  APIFixture: async ({ request }, use) => {
-    customRequest = request;
-    await use(customRequest);
-  },
+  customFixture: [async ({ page, request }, use, testInfo: TestInfo) => {
+    if(testInfo.project.name === "Mailfence") {
+      await Utils.deleteFolder('allure-results');
+      customPage = page;
+      customRequest = request;
+      await page.goto("/");
+      await use("");
+      await page.close();
+      await Utils.deleteFolder('test-results');
+    } else if (testInfo.project.name === "API"){
+      customRequest = request;
+      await use("");
+    }
+  }, {auto: true}]
 });
 
 /**
